+++ src/main/kotlin/com/pestphp/pest/configuration/PestVersionDetector.kt
@@ -1,6 +1,10 @@
 package com.pestphp.pest.configuration
 
+import com.intellij.openapi.project.Project
 import com.jetbrains.php.PhpTestFrameworkVersionDetector
+import com.jetbrains.php.config.interpreters.PhpInterpreter
+import com.jetbrains.php.testFramework.PhpTestFrameworkConfiguration
+import com.jetbrains.php.testFramework.PhpTestFrameworkVersionCache
 import com.pestphp.pest.PestBundle
 import org.jetbrains.annotations.Nls
 
@@ -15,5 +19,20 @@ class PestVersionDetector : PhpTestFrameworkVersionDetector<String>() {
 
     companion object {
         val instance = PestVersionDetector()
+    }
+
+    fun getVersionWithCache(
+        project: Project,
+        interpreter: PhpInterpreter,
+        config: PhpTestFrameworkConfiguration,
+        executable: String?
+    ): String {
+        val cached = PhpTestFrameworkVersionCache.getCache(project, config)
+
+        if (cached.isNotBlank()) {
+            return cached
+        }
+
+        return super.getVersion(project, interpreter, executable)
     }
 }